Improve LMK output (#2)
* Make LMK spammer look up choice by default

Makes the LMK spammer get a list of choices by default, then use that
list to print the actual message of the choice being spammed. This makes
it more obvious that the correct choice is actually being sent.

* Update format of output

Places the timestamp before the choice and removes the `Choice: ` prefix
before the choice content/id.

* Update --no-choice-lookup help

* Update --no-choice-lookup help

* Add custom output for specific error

When an invalid choice ID is passed, the API returns an error saying
`Argument 'question' required`. This is misleading, so we print our own
output explaining the actual error instead.